fixed review comment
